refactor :: navigation package
