+++ app/src/main/java/team/devlib/android/navigation/RootScreen.kt
@@ -1,4 +1,4 @@
-package team.devlib.android
+package team.devlib.android.navigation
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
@@ -8,6 +8,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import team.devlib.android.component.DmsBottomAppBar
+import team.devlib.android.navigation.NavigationRoute
 
 @Composable
 internal fun RootScreen() {
